backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contextlib2==21.6.0
cron_descriptor==2.0.6
cryptography==46.0.5
distro==1.9.0
Django==5.2.11
django-allauth==65.14.3
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-ninja==1.5.3
django-ninja-extra==0.31.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
fido2==2.1.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
injector==0.24.0
jiter==0.13.0
kombu==5.6.2
openai==2.21.0
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
